ui(refactor): split search.js and react.js into multiple files
