+++ scripts/superdesk-search/constants.js
@@ -1,0 +1,7 @@
+export const PARAMETERS = Object.freeze({
+    unique_name: 'Unique Name',
+    original_creator: 'Creator',
+    from_desk: 'From Desk',
+    to_desk: 'To Desk',
+    spike: 'In Spiked'
+});
