+++ scripts/superdesk-search/controllers/index.js
@@ -1,0 +1,1 @@
+export { MultiActionBarController } from './MultiActionBarController';
