+++ scripts/superdesk-search/directives/SaveSearch.js
@@ -1,0 +1,94 @@
+SaveSearch.$inject = ['$location', 'asset', 'api', 'session', 'notify', 'gettext'];
+
+/**
+ * Opens and manages save search panel
+ */
+export function SaveSearch($location, asset, api, session, notify, gettext) {
+    return {
+        templateUrl: asset.templateUrl('superdesk-search/views/save-search.html'),
+        link: function(scope, elem) {
+            scope.edit = null;
+            scope.activateSearchPane = false;
+
+            scope.$on('edit:search', function(event, args)  {
+                scope.activateSearchPane = false;
+                scope.editingSearch = args;
+                scope.edit = _.create(scope.editingSearch) || {};
+            });
+
+            scope.editItem = function () {
+                scope.activateSearchPane = true;
+                scope.edit = _.create(scope.editingSearch) || {};
+            };
+
+            scope.saveas = function() {
+                scope.activateSearchPane = true;
+                scope.edit = _.clone(scope.editingSearch) || {};
+                delete scope.edit._id;
+                scope.edit.name = '';
+                scope.edit.description = '';
+            };
+
+            scope.cancel = function () {
+                scope.sTab = scope.editingSearch ? false : true;
+                scope.resetEditingSearch();
+                scope.edit = null;
+                scope.activateSearchPane = false;
+            };
+
+            scope.clear = function() {
+                scope.resetEditingSearch();
+                scope.edit = null;
+                $location.url($location.path());
+            };
+
+            /**
+             * Patches or posts the given search
+             */
+            scope.save = function(editSearch) {
+
+                function onSuccess() {
+                    notify.success(gettext('Search was saved successfully'));
+                    scope.cancel();
+                    scope.sTab = false;
+                    scope.edit = null;
+                }
+
+                function onFail(error) {
+                    scope.edit = null;
+                    if (angular.isDefined(error.data._message)) {
+                        notify.error(error.data._message);
+                    } else {
+                        notify.error(gettext('Error. Search could not be saved.'));
+                    }
+                }
+
+                var search = getFilters(_.clone($location.search()));
+                editSearch.filter = {query: search};
+                var originalSearch = {};
+
+                if (editSearch._id) {
+                    originalSearch = scope.editingSearch;
+                }
+
+                api('saved_searches', session.identity).save(originalSearch, editSearch).then(onSuccess, onFail);
+            };
+
+            /**
+             * Converts the integer fields: priority and urgency to objects
+             * within a given search
+             *
+             * @return {Object} the updated search object
+             */
+            function getFilters(search) {
+                _.forOwn(search, function(value, key) {
+                    if (_.includes(['priority', 'urgency'], key)) {
+                        search[key] = JSON.parse(value);
+                    }
+                });
+
+                return search;
+            }
+        }
+    };
+}
