+++ scripts/superdesk-search/directives/SearchContainer.js
@@ -1,0 +1,9 @@
+export function SearchContainer() {
+    return {
+        controller: ['$scope', '$location', function SearchContainerController($scope, $location) {
+            this.flags = $scope.flags || {};
+            var query = _.omit($location.search(), '_id');
+            this.flags.facets = !_.isEmpty(query);
+        }]
+    };
+}
