+++ scripts/superdesk-search/services/index.js
@@ -1,0 +1,4 @@
+export { SearchService } from './SearchService';
+export { TagService } from './TagService';
+export { SavedSearchService } from './SavedSearchService';
+export { MonitoringState } from './MonitoringState';
